Rename library / file from BigNum.js to Big.js, Refactor BigInt to support number | string inputs

+++ src/index.ts
@@ -1,1 +1,1 @@
-export {default as BigInt} from './bigInt';
+export {default as Int} from './bigInt';
